# api/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cineca Agentic Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/tenants": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "List tenants",
        "description": "Return a paginated list of tenants registered in the platform.\n\nAdmin-only endpoint that enumerates tenant records. The response is paginated and may be\nETagged for caching. Each tenant includes basic metadata such as id, name, admin_email and\nfree-form `metadata` useful for tenant-scoped configuration.",
        "operationId": "list_tenants_v1_admin_tenants_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  },
                  "title": "Response List Tenants V1 Admin Tenants Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-internal": true
      },
      "post": {
        "tags": [
          "Tenants"
        ],
        "summary": "Create tenant",
        "description": "Create a new tenant record.\n\nAdmin-only operation that provisions a tenant id and basic tenant metadata. For auditing the\nrequest must include the `X-Tenant-Id` header indicating which administrative tenant or\ncontext initiated the creation. Returns the created Tenant object on success.",
        "operationId": "create_tenant_v1_admin_tenants_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "x-internal": true
      }
    },
    "/v1/admin/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "Tenants"
        ],
        "summary": "Get tenant by id",
        "description": "Retrieve a tenant by its id.\n\nAdmin-only read endpoint that returns the tenant record for the requested `tenant_id`.\nReturns HTTP 404 when the tenant cannot be found.",
        "operationId": "get_tenant_v1_admin_tenants__tenant_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "x-internal": true
      },
      "patch": {
        "tags": [
          "Tenants"
        ],
        "summary": "Patch tenant",
        "description": "Update tenant metadata or administrative contact information.\n\nAdmin-only operation that applies a partial update to the tenant record. Supply only the\nfields to change in the request body (name, admin_email, metadata). Returns the updated\ntenant record on success.",
        "operationId": "patch_tenant_v1_admin_tenants__tenant_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "x-internal": true
      },
      "delete": {
        "tags": [
          "Tenants"
        ],
        "summary": "Delete tenant",
        "description": "Delete a tenant from the platform.\n\nAdmin-only destructive operation that removes the tenant record and any associated\ntenant-scoped configuration. Use with caution — callers should ensure dependent resources\nare cleaned up first. Returns a simple `{ \"ok\": true }` on success.",
        "operationId": "delete_tenant_v1_admin_tenants__tenant_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Tenant V1 Admin Tenants  Tenant Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "x-internal": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTenantRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique id for the new tenant"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name for the tenant"
          },
          "admin_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Email",
            "description": "Contact email for tenant admin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CreateTenantRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Tenant": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique tenant identifier"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-friendly tenant name"
          },
          "admin_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Email",
            "description": "Contact email for tenant administrator"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Arbitrary tenant metadata for operator use"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Tenant"
      },
...
      "UpdateTenantRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "New display name for the tenant"
          },
          "admin_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Email",
            "description": "Updated admin contact email"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Partial metadata updates (replace or merge depending on service)"
          }
        },
        "type": "object",
        "title": "UpdateTenantRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          }
        },
        "required": [
          "status"
        ],
        "example": {
          "type": "about:blank",
          "title": "Not Found",
          "status": 404,
          "detail": "Item not found",
          "traceId": "..."
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/v1/auth/token"
          }
        }
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/v1/auth/token",
            "scopes": {
              "admin": "Admin operations",
              "internal": "Internal-only operations",
              "user": "User-level actions"
            }
          },
          "authorizationCode": {
            "authorizationUrl": "/v1/auth/authorize",
            "tokenUrl": "/v1/auth/token",
            "scopes": {
              "openid": "OpenID Connect scope",
              "profile": "Profile info",
              "email": "Email info"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/{version}",
      "variables": {
        "version": {
          "default": "v1",
          "enum": [
            "v1",
            "v2"
          ]
        }
      }
    }
  ],
...
}